changing long to integer

+++ G074HW1.java
@@ -44,7 +44,7 @@ public class G074HW1{
         //Print the number of rows read from the input file
         System.out.println("Number of rows read from the input file = " + rawData.count());
 
-        JavaPairRDD<String, Long> productCustomer;
+        JavaPairRDD<String, Integer> productCustomer;
 
         //Task 2
         productCustomer = rawData
@@ -55,18 +55,18 @@ public class G074HW1{
                     int quantity = Integer.parseInt(tokens[3]);
                     String customerID = tokens[6];
                     String country = tokens[7];
-                    ArrayList<Tuple2<String, Long>> pairs = new ArrayList<>();
+                    ArrayList<Tuple2<String, Integer>> pairs = new ArrayList<>();
                     if((S.compareTo(country) == 0 | S.compareTo("all") == 0) & quantity > 0) { //Selecting the data with country equal to S and quantity > 0
                         //The following passage has been done because we had to avoid the use of distinct, so we had to do a kind of dummy map phase
                         //Create a pair with the key equals to the union of the two ID separated by a "-" : "productID-customerID"
                         //The value of the pair is zero in long format
-                        pairs.add(new Tuple2<>(productID + "-" + customerID, 0L));
+                        pairs.add(new Tuple2<>(productID + "-" + customerID, 0));
                     }
                     return pairs.iterator();
                 }).groupByKey().mapToPair(pair -> {
                     //After grouping the couples that have the same key, so same productID and customerID, We parse the key into productID and customerID
                     String[] tokens = pair._1.split("-");
-                    return new Tuple2<>(tokens[0], Long.parseLong(tokens[1]));
+                    return new Tuple2<>(tokens[0], Integer.parseInt(tokens[1]));
                 });
 
         System.out.println("Number of rows after filtering = " + productCustomer.count());
